requirements.txt
amqp==2.5.2
asgiref==3.2.3
astroid==2.3.3
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.8.4
pylint==2.4.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==2.1.0
